Move authenticators to their own files
