+++ chttp/auth.go
@@ -2,7 +2,6 @@ package chttp
 
 import (
 	"context"
-	"net/http"
 	"net/http/cookiejar"
 
 	"golang.org/x/net/publicsuffix"
@@ -14,77 +13,6 @@ import (
 // Authenticator is an interface that provides authentication to a server.
 type Authenticator interface {
 	Authenticate(context.Context, *Client) error
-}
-
-// BasicAuth provides HTTP Basic Auth for a client.
-type BasicAuth struct {
-	Username string
-	Password string
-
-	// transport stores the original transport that is overridden by this auth
-	// mechanism
-	transport http.RoundTripper
-}
-
-var _ Authenticator = &BasicAuth{}
-
-// RoundTrip fulfills the http.RoundTripper interface. It sets HTTP Basic Auth
-// on outbound requests.
-func (a *BasicAuth) RoundTrip(req *http.Request) (*http.Response, error) {
-	req.SetBasicAuth(a.Username, a.Password)
-	transport := a.transport
-	if transport == nil {
-		transport = http.DefaultTransport
-	}
-	return transport.RoundTrip(req)
-}
-
-// Authenticate sets HTTP Basic Auth headers for the client.
-func (a *BasicAuth) Authenticate(ctx context.Context, c *Client) error {
-	a.transport = c.Transport
-	c.Transport = a
-	return nil
-}
-
-// CookieAuth provides CouchDB Cookie auth services as described at
-// http://docs.couchdb.org/en/2.0.0/api/server/authn.html#cookie-authentication
-//
-// CookieAuth stores authentication state after use, so should not be re-used.
-type CookieAuth struct {
-	Username string `json:"name"`
-	Password string `json:"password"`
-
-	client *Client
-}
-
-var _ Authenticator = &CookieAuth{}
-
-// Authenticate initiates a session with the CouchDB server.
-func (a *CookieAuth) Authenticate(ctx context.Context, c *Client) error {
-	if err := a.setCookieJar(c); err != nil {
-		return err // impossible error
-	}
-	a.client = c
-	opts := &Options{
-		Body: EncodeBody(a),
-	}
-	if _, err := c.DoError(ctx, kivik.MethodPost, "/_session", opts); err != nil {
-		return err
-	}
-	return ValidateAuth(ctx, a.Username, c)
-}
-
-// Cookie returns the current session cookie if found, or nil if not.
-func (a *CookieAuth) Cookie() *http.Cookie {
-	if a.client == nil {
-		return nil
-	}
-	for _, cookie := range a.client.Jar.Cookies(a.client.dsn) {
-		if cookie.Name == kivik.SessionCookieName {
-			return cookie
-		}
-	}
-	return nil
 }
 
 // ValidateAuth validates that the requested username is authenticated.
