+++ chttp/auth_test.go
@@ -3,96 +3,17 @@ package chttp
 import (
 	"context"
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
 	"net/http/cookiejar"
 	"net/http/httptest"
 	"net/url"
-	"strings"
 	"testing"
 
-	"github.com/flimzy/diff"
 	"github.com/flimzy/testy"
 	"golang.org/x/net/publicsuffix"
 
 	"github.com/go-kivik/kivik"
 )
-
-func TestBasicAuthRoundTrip(t *testing.T) {
-	type rtTest struct {
-		name     string
-		auth     *BasicAuth
-		req      *http.Request
-		expected *http.Response
-		cleanup  func()
-	}
-	tests := []rtTest{
-		{
-			name: "Provided transport",
-			req:  httptest.NewRequest("GET", "/", nil),
-			auth: &BasicAuth{
-				Username: "foo",
-				Password: "bar",
-				transport: customTransport(func(req *http.Request) (*http.Response, error) {
-					u, p, ok := req.BasicAuth()
-					if !ok {
-						t.Error("BasicAuth not set in request")
-					}
-					if u != "foo" || p != "bar" {
-						t.Errorf("Unexpected user/password: %s/%s", u, p)
-					}
-					return &http.Response{StatusCode: 200}, nil
-				}),
-			},
-			expected: &http.Response{StatusCode: 200},
-		},
-		func() rtTest {
-			h := func(w http.ResponseWriter, r *http.Request) {
-				u, p, ok := r.BasicAuth()
-				if !ok {
-					t.Error("BasicAuth not set in request")
-				}
-				if u != "foo" || p != "bar" {
-					t.Errorf("Unexpected user/password: %s/%s", u, p)
-				}
-				w.Header().Set("Date", "Wed, 01 Nov 2017 19:32:41 GMT")
-				w.Header().Set("Content-Type", "application/json")
-			}
-			s := httptest.NewServer(http.HandlerFunc(h))
-			return rtTest{
-				name: "default transport",
-				auth: &BasicAuth{Username: "foo", Password: "bar"},
-				req:  httptest.NewRequest("GET", s.URL, nil),
-				expected: &http.Response{
-					Status:     "200 OK",
-					StatusCode: 200,
-					Proto:      "HTTP/1.1",
-					ProtoMajor: 1,
-					ProtoMinor: 1,
-					Header: http.Header{
-						"Content-Length": {"0"},
-						"Content-Type":   {"application/json"},
-						"Date":           {"Wed, 01 Nov 2017 19:32:41 GMT"},
-					},
-				},
-				cleanup: func() { s.Close() },
-			}
-		}(),
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			res, err := test.auth.RoundTrip(test.req)
-			if err != nil {
-				t.Fatal(err)
-			}
-			res.Body = nil
-			res.Request = nil
-			if d := diff.Interface(test.expected, res); d != nil {
-				t.Error(d)
-			}
-		})
-	}
-}
 
 type mockRT struct {
 	resp *http.Response
@@ -215,154 +136,4 @@ func TestAuthenticate(t *testing.T) {
 		_, err = c.DoError(ctx, "GET", "/foo", nil)
 		testy.StatusError(t, test.err, test.status, err)
 	})
-}
-
-func TestCookieAuthAuthenticate(t *testing.T) {
-	tests := []struct {
-		name           string
-		auth           *CookieAuth
-		client         *Client
-		status         int
-		err            string
-		expectedCookie *http.Cookie
-	}{
-		{
-			name: "standard request",
-			auth: &CookieAuth{
-				Username: "foo",
-				Password: "bar",
-			},
-			client: &Client{
-				Client: &http.Client{
-					Transport: &mockRT{
-						resp: &http.Response{
-							Header: http.Header{
-								"Set-Cookie": []string{
-									"AuthSession=cm9vdDo1MEJCRkYwMjq0LO0ylOIwShrgt8y-UkhI-c6BGw; Version=1; Path=/; HttpOnly",
-								},
-							},
-							Body: ioutil.NopCloser(strings.NewReader(`{"userCtx":{"name":"foo"}}`)),
-						},
-					},
-				},
-				dsn: &url.URL{Scheme: "http", Host: "foo.com"},
-			},
-			expectedCookie: &http.Cookie{
-				Name:  kivik.SessionCookieName,
-				Value: "cm9vdDo1MEJCRkYwMjq0LO0ylOIwShrgt8y-UkhI-c6BGw",
-			},
-		},
-		{
-			name: "Invalid JSON response",
-			auth: &CookieAuth{
-				Username: "foo",
-				Password: "bar",
-			},
-			client: &Client{
-				Client: &http.Client{
-					Jar: &cookiejar.Jar{},
-					Transport: &mockRT{
-						resp: &http.Response{
-							Body: ioutil.NopCloser(strings.NewReader(`{"asdf"}`)),
-						},
-					},
-				},
-				dsn: &url.URL{Scheme: "http", Host: "foo.com"},
-			},
-			status: kivik.StatusBadResponse,
-			err:    "invalid character '}' after object key",
-		},
-		{
-			name: "names don't match",
-			auth: &CookieAuth{
-				Username: "foo",
-				Password: "bar",
-			},
-			client: &Client{
-				Client: &http.Client{
-					Transport: &mockRT{
-						resp: &http.Response{
-							Header: http.Header{
-								"Set-Cookie": []string{
-									"AuthSession=cm9vdDo1MEJCRkYwMjq0LO0ylOIwShrgt8y-UkhI-c6BGw; Version=1; Path=/; HttpOnly",
-								},
-							},
-							Body: ioutil.NopCloser(strings.NewReader(`{"userCtx":{"name":"notfoo"}}`)),
-						},
-					},
-				},
-				dsn: &url.URL{Scheme: "http", Host: "foo.com"},
-			},
-			status: kivik.StatusBadResponse,
-			err:    "auth response for unexpected user",
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			err := test.auth.Authenticate(context.Background(), test.client)
-			testy.StatusError(t, test.err, test.status, err)
-			cookie := test.auth.Cookie()
-			if d := diff.Interface(test.expectedCookie, cookie); d != nil {
-				t.Error(d)
-			}
-		})
-	}
-}
-
-func TestCookie(t *testing.T) {
-	tests := []struct {
-		name     string
-		auth     *CookieAuth
-		expected *http.Cookie
-	}{
-		{
-			name:     "No cookie jar",
-			auth:     &CookieAuth{},
-			expected: nil,
-		},
-		{
-			name:     "No dsn",
-			auth:     &CookieAuth{},
-			expected: nil,
-		},
-		{
-			name:     "no cookies",
-			auth:     &CookieAuth{},
-			expected: nil,
-		},
-		{
-			name: "cookie found",
-			auth: func() *CookieAuth {
-				dsn, err := url.Parse("http://example.com/")
-				if err != nil {
-					t.Fatal(err)
-				}
-				jar, err := cookiejar.New(&cookiejar.Options{PublicSuffixList: publicsuffix.List})
-				if err != nil {
-					t.Fatal(err)
-				}
-				jar.SetCookies(dsn, []*http.Cookie{
-					{Name: kivik.SessionCookieName, Value: "foo"},
-					{Name: "other", Value: "bar"},
-				})
-				return &CookieAuth{
-					client: &Client{
-						dsn: dsn,
-						Client: &http.Client{
-							Jar: jar,
-						},
-					},
-				}
-			}(),
-			expected: &http.Cookie{Name: kivik.SessionCookieName, Value: "foo"},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			result := test.auth.Cookie()
-			if d := diff.Interface(test.expected, result); d != nil {
-				t.Error(d)
-			}
-		})
-	}
 }
