+++ chttp/basicauth.go
@@ -1,0 +1,36 @@
+package chttp
+
+import (
+	"context"
+	"net/http"
+)
+
+// BasicAuth provides HTTP Basic Auth for a client.
+type BasicAuth struct {
+	Username string
+	Password string
+
+	// transport stores the original transport that is overridden by this auth
+	// mechanism
+	transport http.RoundTripper
+}
+
+var _ Authenticator = &BasicAuth{}
+
+// RoundTrip fulfills the http.RoundTripper interface. It sets HTTP Basic Auth
+// on outbound requests.
+func (a *BasicAuth) RoundTrip(req *http.Request) (*http.Response, error) {
+	req.SetBasicAuth(a.Username, a.Password)
+	transport := a.transport
+	if transport == nil {
+		transport = http.DefaultTransport
+	}
+	return transport.RoundTrip(req)
+}
+
+// Authenticate sets HTTP Basic Auth headers for the client.
+func (a *BasicAuth) Authenticate(_ context.Context, c *Client) error {
+	a.transport = c.Transport
+	c.Transport = a
+	return nil
+}
